requirements.txt
# Supabase client
supabase>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# API framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Other dependencies (add as needed)
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn
